+++ packages/plugin-packages/mashroom-portal-remote-app-registry-k8s/src/js/context.ts
@@ -4,6 +4,8 @@ import KubernetesServiceRegistry from './registry/KubernetesServiceRegistry';
 import type {Context} from '../../type-definitions';
 
 export const _registry = new KubernetesServiceRegistry();
+let _namespaces: Array<string> = [];
+let _serviceLabelSelector: string | null = null;
 let _serviceNameFilter = '';
 let _error: string | null = null;
 let _lastScan = 0;
@@ -24,6 +26,18 @@ const context: Context = {
     },
     set lastScan(lastScan: number) {
         _lastScan = lastScan;
+    },
+    get namespaces() {
+        return _namespaces;
+    },
+    set namespaces(namespaces: Array<string>) {
+        _namespaces = namespaces;
+    },
+    get serviceLabelSelector() {
+        return _serviceLabelSelector;
+    },
+    set serviceLabelSelector(serviceLabelSelector: string | null) {
+        _serviceLabelSelector = serviceLabelSelector;
     },
     get serviceNameFilter() {
         return _serviceNameFilter;